Moved wink scenes from switches to scenes
